casts/changes to cope with size_t vs vprintf
svn:revision:42889
svn:branch:trunk
svn:account:erikgreenwald

+++ src/conv/bot_shell-vtk.c
@@ -571,7 +571,7 @@ main(int argc, char *argv[])
 	database_index++;
     }
     fprintf( fd_out, "\nASCII\n\nDATASET POLYDATA\n" );
-    fprintf( fd_out, "POINTS %ld float\n", verts->curr_vert );
+    fprintf( fd_out, "POINTS %ld float\n", (long int)verts->curr_vert );
     for ( i=0; i<(size_t)verts->curr_vert; i++ ) {
 	if ( use_normals ) {
 	    fprintf( fd_out, "%g %g %g\n", V3ARGS( &verts->the_array[i*6] ) );
@@ -584,7 +584,7 @@ main(int argc, char *argv[])
 	fprintf( fd_out, "3 %ld %ld %ld\n", V3ARGS( &faces[i*3] ) );
     }
     if ( use_normals ) {
-	fprintf( fd_out, "POINT_DATA %ld\n", verts->curr_vert );
+	fprintf( fd_out, "POINT_DATA %ld\n", (long int)verts->curr_vert );
 	fprintf( fd_out, "NORMALS default float\n" );
 	for ( i=0; i<(size_t)verts->curr_vert; i++ ) {
 	    fprintf( fd_out, "%g %g %g\n", V3ARGS( &verts->the_array[i*6+3] ) );
